Cambios generales

+++ resources/views/bombero/lista.blade.php
@@ -10,13 +10,9 @@
 
     <div class="col-sm-12">
       <div class="col-md-7 col-sm-12 text-right" style="padding-top: 20px;">
-        @php
-          $jerarquias[0]="Jerarquia";
-          $jerarquias=array_merge($jerarquias, config('selects.jerarquia'));
-        @endphp
         {{Form::model(Request::all(),['route' => 'bombero.index', 'class' => 'form-horizontal', 'method' => 'GET'])}}
           <div class="col-sm-3">
-            {{Form::select('jerarquia', $jerarquias,0, ['class' => 'form-control'])}}
+            {{Form::select('jerarquia', [0=>'Jerarquia'] + config('selects.jerarquia'),0, ['class' => 'form-control'])}}
           </div>
           <div class="col-sm-4">
             {{Form::text('nombre', null, ['placeholder'=>"Buscar por Apellido/Nombre", 'class' => 'form-control'])}}
@@ -57,15 +53,21 @@
               <td>{{$bombero->telefono}}</td>
               <td>{{\Carbon\Carbon::parse($bombero->fecha_nacimiento)->format('d/m/Y')}}</td>
               @if (Auth::user()->admin)
-                <td><a href="{{ route('bombero.edit', $bombero->id) }}"><button class="glyphicon glyphicon-edit"></button></a></td>
                 <td>
-                  {{ Form::open(['route' => ['bombero.destroy', $bombero->id], 'method' => 'DELETE']) }}
-                      <button type="submit" class="glyphicon glyphicon-trash"></button>
-                  {{ Form::close() }}
+                  <a href="{{ route('bombero.edit', $bombero->id) }}"><button class="glyphicon glyphicon-edit"></button></a></td>
+                <td>
+                  @if (count($bombero->servicios)==0)
+                    {{ Form::open(['route' => ['bombero.destroy', $bombero->id], 'method' => 'DELETE']) }}
+                        <button type="submit" class="glyphicon glyphicon-trash"></button>
+                    {{ Form::close() }}
+                  @else
+                    <button type="submit" class="glyphicon glyphicon-ban-circle" title="Imposible eliminar, participo en al menos un servicio"></button>
+
+                  @endif
                 </td>
               @else
                 <td colspan="2">
-                  <button type="submit" class="btn glyphicon glyphicon-ban-circle ban" title="Sin permisos para eliminar/modificar"></button>
+                  <button type="submit" class="glyphicon glyphicon-ban-circle" title="Sin permisos para eliminar/modificar"></button>
                 </td>
               @endif
             </tr>
@@ -74,7 +76,7 @@
           <br>
         </table>
         <div class="text-center">
-          {{ $bomberos->appends(Request::only(['legajo']))->links()}}
+          {{ $bomberos->appends(Request::all())->links()}}
         </div>
     </div>
   </div>
